+++ Profiler.h
@@ -18,7 +18,7 @@ public:
 	void WriteToProfiler(const std::string& _inputStr);
 	void WriteToProfiler(const std::string& _inputStr, float _value);
 	void WriteToProfiler(const std::string& _inputStr, double _value);
-	void WriteToProfiler(const std::string& _inputStr, int _value);
+	void WriteToProfiler(const std::string& _inputStr, uint64_t _value);
 
 private:
 	Profiler();
